Major changes: redefine purchase process

+++ accounting_new/api.py
@@ -10,65 +10,66 @@ from rest_framework.response import Response
 from nakhll.authentications import CsrfExemptSessionAuthentication
 from cart.managers import CartManager
 from logistic.interfaces import PostPriceSettingInterface
+from payoff.exceptions import NoAddressException, InvoiceExpiredException,\
+            InvalidInvoiceStatusException, OutOfPostRangeProductsException
 from .models import Invoice
-from .serializers import InvoiceWriteSerializer, InvoiceReadSerializer
+from .serializers import InvoiceWriteSerializer, InvoiceRetrieveSerializer
 from .permissions import IsInvoiceOwner
 
 
 class InvoiceViewSet(viewsets.GenericViewSet, mixins.CreateModelMixin, 
                     mixins.UpdateModelMixin, mixins.RetrieveModelMixin,
-                    mixins.DestroyModelMixin):
+                    mixins.DestroyModelMixin, mixins.ListModelMixin):
     permission_classes = [IsInvoiceOwner, permissions.IsAuthenticated, ]
     authentication_classes = [CsrfExemptSessionAuthentication, ]
     queryset = Invoice.objects.all()
 
+    def get_queryset(self):
+        return super().get_queryset().filter(user=self.request.user)
 
     def get_serializer_class(self):
-        if self.action in ['list', 'retrieve']:
-            return InvoiceReadSerializer
+        if self.action == 'list':
+            return InvoiceRetrieveSerializer
+        elif self.action == 'retrieve':
+            return InvoiceRetrieveSerializer
         else:
             return InvoiceWriteSerializer
 
-    def create_invoice(self, request):
+    def __create_invoice(self, request):
         active_cart = CartManager.user_active_cart(request.user)
-        data = {'cart': active_cart.id }
-        serializer = self.get_serializer(data=data)
-        serializer.is_valid(raise_exception=True)
-        return self.perform_create(serializer)
-
-
-
-    def get_object(self):
-        # TODO: System must support multi invoice per user (more on clickup#1fu8awk)
-        active_cart = CartManager.user_active_cart(self.request.user)
-        invoice = Invoice.objects.filter(cart=active_cart).first() or self.create_invoice(self.request)
-        # if invoice.status == Invoice.Statuses.PAYING:
-            # raise ValidationError('فاکتور شما در حال پرداخت می‌باشد و امکان دسترسی به آن وجود ندارد')
-        invoice.status = Invoice.Statuses.AWAIT_PAYING
-        invoice.save()
+        if not active_cart.items.all():
+            raise ValidationError('سبد خرید خالی است')
+        invoice = active_cart.convert_to_invoice()
         return invoice
 
+    # def get_object(self):
+    #     active_cart = CartManager.user_active_cart(self.request.user)
+    #     invoice = Invoice.objects.filter(cart=active_cart).first() or self.create_invoice(self.request)
+    #     invoice.status = Invoice.Statuses.AWAIT_PAYMENT
+    #     invoice.save()
+    #     return invoice
+
     def create(self, request, *args, **kwargs):
-        ''' each user can only have one invoice with status of completing '''
-        invoice = self.get_object()
-        serializer = InvoiceReadSerializer(invoice)
+        ''' each user can have many invoices '''
+        try:
+            invoice = self.__create_invoice(request)
+            #TODO: Create alert
+        except ValidationError as ex:
+            return Response({'detail': f'خطا در ساخت فاکتور: {ex}'}, status=status.HTTP_400_BAD_REQUEST)
+        serializer = InvoiceRetrieveSerializer(invoice)
         headers = self.get_success_headers(serializer.data)
         return Response(serializer.data, status=status.HTTP_200_OK, headers=headers)
 
-    def perform_create(self, serializer):
-        invoice = serializer.save()
-        #TODO: Create alert
-        return invoice
 
 
-    @action(methods=['GET'], detail=False)
-    def active_invoice(self, request):
-        invoice = self.get_object()
-        serializer = InvoiceReadSerializer(invoice)
-        return Response(serializer.data, status=status.HTTP_200_OK)
+    # @action(methods=['GET'], detail=False)
+    # def active_invoice(self, request):
+    #     invoice = self.get_object()
+    #     serializer = InvoiceReadSerializer(invoice)
+    #     return Response(serializer.data, status=status.HTTP_200_OK)
 
-    @action(methods=['PATCH'], detail=False)
-    def set_coupon(self, request):
+    @action(methods=['PATCH'], detail=True)
+    def set_coupon(self, request, pk):
         ''' Verify and calculate user coupon and return discount amount
         
             Get invoice with pk as id, insure that invoice belogs to user
@@ -87,8 +88,8 @@ class InvoiceViewSet(viewsets.GenericViewSet, mixins.CreateModelMixin,
             coupon.apply(invoice)
         return Response({'coupon': coupon.code, 'result': coupon.final_price, 'errors': coupon.errors}, status=status.HTTP_200_OK)
 
-    @action(methods=['PATCH'], detail=False)
-    def unset_coupon(self, request):
+    @action(methods=['PATCH'], detail=True)
+    def unset_coupon(self, request, pk):
         ''' Unset coupon from invoice
         
             Get invoice with pk as id, insure that invoice belogs to user
@@ -103,12 +104,14 @@ class InvoiceViewSet(viewsets.GenericViewSet, mixins.CreateModelMixin,
         serializer.is_valid(raise_exception=True)
         coupon = serializer.validated_data.get('coupon')
         coupon_usage = invoice.coupon_usages.filter(coupon=coupon).first()
+        if not coupon_usage:
+            return Response({'result': 'چنین کوپنی برای این فاکتور ثبت نشده است'}, status=status.HTTP_400_BAD_REQUEST)
         coupon_usage.delete()
         serializer.save()
         return Response({'result': 0}, status=status.HTTP_200_OK)
     
-    @action(methods=['PATCH'], detail=False)
-    def set_address(self, request):
+    @action(methods=['PATCH'], detail=True)
+    def set_address(self, request, pk):
         ''' Calculate logistic price from shops to user address 
         
             Get invoice and insure that it is for requested user. Also it 
@@ -122,15 +125,17 @@ class InvoiceViewSet(viewsets.GenericViewSet, mixins.CreateModelMixin,
         invoice = self.get_object()
         serializer = InvoiceWriteSerializer(instance=invoice, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
-        serializer.save(cart=invoice.cart)
-        logistic, is_created = PostPriceSetting.objects.get_or_create()
-        out_of_range_shops = logistic.get_out_of_range_products(invoice)
-        post_price = logistic.get_post_price(invoice)
-        return Response({'post_price': post_price, 'out_of_range': out_of_range_shops}, status=status.HTTP_200_OK)
+        serializer.save()
+        # logistic, is_created = PostPriceSetting.objects.get_or_create()
+        # out_of_range_shops = logistic.get_out_of_range_products(invoice)
+        # post_price = logistic.get_post_price(invoice)
+        post_price = invoice.logistic_price
+        out_of_range_products = invoice.logistic_errors
+        return Response({'post_price': post_price, 'out_of_range': out_of_range_products}, status=status.HTTP_200_OK)
 
 
-    @action(methods=['GET'], detail=False)
-    def pay(self, request):
+    @action(methods=['GET'], detail=True)
+    def pay(self, request, pk):
         ''' Get an invoice and send it to payment app
         
             Request for invoice should came from owner.
@@ -144,24 +149,33 @@ class InvoiceViewSet(viewsets.GenericViewSet, mixins.CreateModelMixin,
             A celery should check for invocies with paying status every hour.
             Invoice should sent to payment status to initiate payment
         '''
-        # TODO: editing items in active cart when cart sent to accounting should be denied
-        # TODO: gettings diffrences is not implemented completely
         invoice = self.get_object()
-        is_differ = invoice.cart.get_diffrences
-        if is_differ:
-            
-            return Response({'error': 'تغییراتی در سبد خرید شما به وجود آمده است. لطفا سبد خرید را بررسی کنید'}, status=status.HTTP_400_BAD_REQUEST)
-        invoice.send_to_payment()
+        # is_differ = invoice.cart.get_diffrences
+        # if is_differ:
+            # return Response({'error': 'تغییراتی در سبد خرید شما به وجود آمده است. لطفا سبد خرید را بررسی کنید'}, status=status.HTTP_400_BAD_REQUEST)
+        try:
+            invoice.send_to_payment()
+        except NoAddressException:
+            return Response({'error': 'آدرس خریدار را تکمیل کنید'}, status=status.HTTP_400_BAD_REQUEST)
+        except InvoiceExpiredException:
+            return Response({'error': 'فاکتور منقضی شده است'}, status=status.HTTP_400_BAD_REQUEST)
+        except InvalidInvoiceStatusException:
+            return Response({'error': 'فاکتور در حال حاضر قابل پرداخت نیست'}, status=status.HTTP_400_BAD_REQUEST)
+        except OutOfPostRangeProductsException as ex:
+            return Response({'error': f'این محصولات خارج از محدوده ارسال شما هستند: {ex}'}, status=status.HTTP_400_BAD_REQUEST)
+        except Exception as ex:
+            return Response({'error': str(ex)}, status=status.HTTP_400_BAD_REQUEST)
+        
 
-    @action(methods=['GET'], detail=False)
-    def confirm_changes(self, request):
-        ''' User confirms for changes made to product and update product_last_state '''
-        invoice = self.get_object()
-        cart = invoice.cart
-        for item in cart.items.all():
-            item.product_last_state = ProductLastStateSerializer(item.product).data
-            item.save()
-        return Response({'result': 'success'}, status=status.HTTP_200_OK)
+    # @action(methods=['GET'], detail=False)
+    # def confirm_changes(self, request):
+    #     ''' User confirms for changes made to product and update product_last_state '''
+    #     invoice = self.get_object()
+    #     cart = invoice.cart
+    #     for item in cart.items.all():
+    #         item.product_last_state = ProductLastStateSerializer(item.product).data
+    #         item.save()
+    #     return Response({'result': 'success'}, status=status.HTTP_200_OK)
